transform/.base/requirements.txt
pandas==1.5.2
python-decouple==3.6
SQLAlchemy==2.0.3
